+++ src/main/scala/edu/ie3/simona/io/result/ResultSinkType.scala
@@ -8,6 +8,8 @@ package edu.ie3.simona.io.result
 
 import edu.ie3.simona.config.SimonaConfig
 
+import java.util.UUID
+
 /** Enumeration to describe all eligible types of
   * [[edu.ie3.datamodel.models.result.ResultEntity]] sink
   */
@@ -24,11 +26,20 @@ object ResultSinkType {
   final case class InfluxDb1x(url: String, database: String, scenario: String)
       extends ResultSinkType
 
+  final case class Kafka(
+      topicNodeRes: String,
+      runId: UUID,
+      bootstrapServers: String,
+      schemaRegistryUrl: String,
+      linger: Int
+  ) extends ResultSinkType
+
   def apply(
       sinkConfig: SimonaConfig.Simona.Output.Sink,
       runName: String
   ): ResultSinkType = {
-    val sink: Seq[Any] = Seq(sinkConfig.csv, sinkConfig.influxDb1x).flatten
+    val sink: Seq[Any] =
+      Seq(sinkConfig.csv, sinkConfig.influxDb1x, sinkConfig.kafka).flatten
 
     if (sink.size > 1)
       throw new IllegalArgumentException(
@@ -40,6 +51,14 @@ object ResultSinkType {
         Csv(params.fileFormat, params.filePrefix, params.fileSuffix)
       case Some(params: SimonaConfig.Simona.Output.Sink.InfluxDb1x) =>
         InfluxDb1x(buildInfluxDb1xUrl(params), params.database, runName)
+      case Some(params: SimonaConfig.ResultKafkaParams) =>
+        Kafka(
+          params.topicNodeRes,
+          UUID.fromString(params.runId),
+          params.bootstrapServers,
+          params.schemaRegistryUrl,
+          params.linger
+        )
       case None =>
         throw new IllegalArgumentException(
           s"No sinks defined! Cannot determine the sink type!"