+++ src/main/scala/edu/ie3/simona/model/participant/load/profile/ProfileLoadModel.scala
@@ -6,8 +6,8 @@
 
 package edu.ie3.simona.model.participant.load.profile
 
-import edu.ie3.datamodel.models.input.system.LoadInput
 import edu.ie3.datamodel.models.profile.StandardLoadProfile
+import edu.ie3.datamodel.models.input.system.LoadInput
 import edu.ie3.simona.model.participant.CalcRelevantData.LoadRelevantData
 import edu.ie3.simona.model.participant.control.QControl
 import edu.ie3.simona.model.participant.load.LoadReference._
@@ -142,7 +142,9 @@ case object ProfileLoadModel {
 
     case LoadReference.EnergyConsumption(energyConsumption) =>
       val loadProfileMax =
-        LoadProfileStore().maxPower(input.getLoadProfile)
+        LoadProfileStore().maxPower(
+          input.getLoadProfile.asInstanceOf[StandardLoadProfile]
+        )
       val sRatedEnergy = LoadModel.scaleSRatedEnergy(
         input,
         energyConsumption,