+++ src/test/groovy/edu/ie3/simona/model/participant/load/ProfileLoadModelTest.groovy
@@ -11,7 +11,6 @@ import edu.ie3.datamodel.models.input.NodeInput
 import edu.ie3.datamodel.models.input.OperatorInput
 import edu.ie3.datamodel.models.input.system.LoadInput
 import edu.ie3.datamodel.models.input.system.characteristic.CosPhiFixed
-import edu.ie3.datamodel.models.profile.StandardLoadProfile
 import edu.ie3.datamodel.models.voltagelevels.GermanVoltageLevelUtils
 import edu.ie3.simona.model.SystemComponent
 import edu.ie3.simona.model.participant.control.QControl
@@ -22,7 +21,6 @@ import tech.units.indriya.ComparableQuantity
 import tech.units.indriya.quantity.Quantities
 
 import javax.measure.quantity.Energy
-import javax.measure.quantity.Power
 import java.time.temporal.ChronoUnit
 import java.util.stream.Collectors
 
@@ -70,11 +68,10 @@ class ProfileLoadModelTest extends Specification {
 	)
 	def testingTolerance = 1e-6 // Equals to 1 W power
 
-	def "A profile load model should be instantiated from valid input correctly"(StandardLoadProfile profile, LoadReference reference, ComparableQuantity<Power> expectedsRated) {
+	def "A profile load model should be instantiated from valid input correctly"() {
 		when:
-		def adaptedInput = loadInput.copy().loadprofile(profile).build()
 		def actual = ProfileLoadModel.apply(
-				adaptedInput,
+				loadInput.copy().loadprofile(profile).build(),
 				foreSeenOperationInterval,
 				1.0,
 				reference)
@@ -83,13 +80,13 @@ class ProfileLoadModelTest extends Specification {
 		abs((actual.sRated() * actual.cosPhiRated()).subtract(expectedsRated).to(MEGAWATT).value.doubleValue()) < testingTolerance
 
 		where:
-		profile                   | reference                                                          || expectedsRated
-		H0 as StandardLoadProfile | new ActivePower(Quantities.getQuantity(268.6, WATT))               || Quantities.getQuantity(268.6, WATT)
-		H0 as StandardLoadProfile | new EnergyConsumption(Quantities.getQuantity(3000d, KILOWATTHOUR)) || Quantities.getQuantity(805.8089, WATT)
-		L0 as StandardLoadProfile | new ActivePower(Quantities.getQuantity(268.6, WATT))               || Quantities.getQuantity(268.6, WATT)
-		L0 as StandardLoadProfile | new EnergyConsumption(Quantities.getQuantity(3000d, KILOWATTHOUR)) || Quantities.getQuantity(721.2, WATT)
-		G0 as StandardLoadProfile | new ActivePower(Quantities.getQuantity(268.6, WATT))               || Quantities.getQuantity(268.6, WATT)
-		G0 as StandardLoadProfile | new EnergyConsumption(Quantities.getQuantity(3000d, KILOWATTHOUR)) || Quantities.getQuantity(721.2, WATT)
+		profile | reference                                                          || expectedsRated
+		H0      | new ActivePower(Quantities.getQuantity(268.6, WATT))               || Quantities.getQuantity(268.6, WATT)
+		H0      | new EnergyConsumption(Quantities.getQuantity(3000d, KILOWATTHOUR)) || Quantities.getQuantity(805.8089, WATT)
+		L0      | new ActivePower(Quantities.getQuantity(268.6, WATT))               || Quantities.getQuantity(268.6, WATT)
+		L0      | new EnergyConsumption(Quantities.getQuantity(3000d, KILOWATTHOUR)) || Quantities.getQuantity(721.2, WATT)
+		G0      | new ActivePower(Quantities.getQuantity(268.6, WATT))               || Quantities.getQuantity(268.6, WATT)
+		G0      | new EnergyConsumption(Quantities.getQuantity(3000d, KILOWATTHOUR)) || Quantities.getQuantity(721.2, WATT)
 	}
 
 	def "A profile load model should reach the targeted maximum power within a year"() {